FIx typos and correct naming conventions
It is difficult for the document reader to understand `./src/api/[api-name]/content-types/restaurant/lifecycles.js` because he might be working on a local codebase and having `restaurant` in the documentation make it confusing. I replaced it with the `[api-name]` so, the reader would know how to define and name that file correctly.