Update admin-customization.md for plugin translation
Add instructions to admin-customization.md for extending a plugin's translations, which requires prefixing the key/value pair (declared in the plugin's translation file) with the plugin's name, for example `"content-type-builder.plugin.name": "Constructeur de Type-Contenu"` in order to translate the `plugin.name` key declared in `content-type-builder` plugin.
This wasn't made clear by the documentation previously, since even for default plugins (such as content-type-builder) some fields are not translatable with the default method because they are declared independently in the plugin's files. I couldn't find any reference in the docs to the fact the plugin's name needed to be added as a prefix in order to extend a plugin's key.