Postgres database user is granted schema privileges
When connecting Strapi to a postgres database, the database user requires public SCHEMA permissions. This can be confusing for the developer since when the permission is not configured, Strapi only throws a 500 error without specific details.

This documentation aims to help postgres/strapi devs avoid friction when setting up new postgres dbs.