Make separate section for Nginx config on droplet
There are some additional notes that need to be included for correct configuration.
Remove the text about admin user as it's not yet possible to access the dashboard from the server.
Also remove tip and Nginx config note at end of setup as it's all located in one section now.